Refactor the blueprint test helper `enableModuleUnification`
