+++ node-tests/blueprints/controller-test.js
@@ -12,9 +12,11 @@ const chai = require('ember-cli-blueprint-test-helpers/chai');
 const expect = chai.expect;
 
 const generateFakePackageManifest = require('../helpers/generate-fake-package-manifest');
-const enableModuleUnification = require('../helpers/module-unification').enableModuleUnification;
-const enableOctane = require('../helpers/setup-test-environment').enableOctane;
 const fixture = require('../helpers/fixture');
+
+const setupTestEnvironment = require('../helpers/setup-test-environment');
+const enableModuleUnification = setupTestEnvironment.enableModuleUnification;
+const enableOctane = setupTestEnvironment.enableOctane;
 
 describe('Blueprint: controller', function() {
   setupTestHooks(this);
