--- node-tests/helpers/module-unification.js
@@ -1,13 +1,0 @@
-function enableModuleUnification() {
-  beforeEach(function() {
-    process.env.EMBER_CLI_MODULE_UNIFICATION = 'true';
-  });
-
-  afterEach(function() {
-    delete process.env.EMBER_CLI_MODULE_UNIFICATION;
-  });
-}
-
-module.exports = {
-  enableModuleUnification,
-};
